Extract json_selection/apply_to.rs from parser.rs.
