+++ apollo-federation/src/sources/connect/json_selection/mod.rs
@@ -1,4 +1,6 @@
+mod apply_to;
 mod helpers;
 mod parser;
 
+pub use apply_to::*;
 pub use parser::*;
