+++ apollo-federation/src/sources/connect/mod.rs
@@ -1,2 +1,8 @@
 mod json_selection;
 mod url_path_template;
+
+pub use json_selection::ApplyTo;
+pub use json_selection::ApplyToError;
+pub use json_selection::Selection;
+pub use json_selection::SubSelection;
+pub use url_path_template::URLPathTemplate;
